backend/requirements.txt
anthropic>=0.40.0
openai>=1.55.0
pydantic>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
rich>=13.9.0
beautifulsoup4>=4.12.0
ddgs>=9.14.0
mcp>=1.27.0
tenacity>=9.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
